CSE368_Project/requirements.txt
pymongo
Flask
Flask-SQLAlchemy
google-genai
dotenv